[util] factorize some methods from scripts to util.py
