+++ scripts/mlp/centroidal/topt.py
@@ -66,39 +66,7 @@ def isNewPhase(tp,k0,k1,cs_com,cs_initGuess,p_id):
         return isNewPhaseFromCS(cs_com,cs_initGuess,p_id)
     else : 
         return isNewPhaseFromContact(tp,k0,k1)
-
-
-def connectFinalPhase(phase,duration = cfg.DURATION_CONNECT_GOAL):
-    if duration <= 0.:
-        return
-    init_state = phase.state_trajectory[-1]
-    final_state = phase.final_state
-    init_control = phase.control_trajectory[-1]
-    t_init = phase.time_trajectory[-1]
-    t_end = t_init + duration
-    com_traj = genCOMTrajFromPhaseStates(t_init,t_end,init_state,final_state,init_control)
-    am_traj = genAMTrajFromPhaseStates(t_init,t_end,init_state,final_state,init_control)
-    i = len(phase.time_trajectory)
-    
-    dt = cfg.SOLVER_DT
-    t = t_init + dt
-    while t < t_end + dt/2. :
-        if t > t_end: # may happen due to numerical imprecision
-            t = t_end
-        c,dc,ddc = com_traj(t)
-        L,dL = am_traj(t)
-        state = np.matrix(np.zeros(9)).T
-        control= np.matrix(np.zeros(6)).T
-        state[0:3] = c
-        state[3:6] = dc
-        control[0:3] = ddc
-        state[6:9] = L
-        control[3:6] = dL
-        phase.state_trajectory.append(state)
-        phase.control_trajectory.append(control)
-        phase.time_trajectory.append(t)
-        t += dt
-           
+ 
 def fillCSFromTimeopt(cs,cs_initGuess,tp):
     cs_com = ContactSequenceHumanoid(cs)
     
@@ -148,7 +116,7 @@ def fillCSFromTimeopt(cs,cs_initGuess,tp):
         k_id +=1
     # timeopt solution is not guarantee to end at the desired final state.
     # so we add a final phase here, with a smooth motion from the final state atteined by timeopt to the desired one      
-    connectFinalPhase(cs_com.contact_phases[-1])
+    connectPhaseTrajToFinalState(cs_com.contact_phases[-1],cfg.DURATION_CONNECT_GOAL)
     return cs_com
         
 # helper method to make the link between timeopt.EndEffector and cs.Patch
